moved SetSortBy method from dbContext into repository specific methods to avoid failed queries.
